# GEOS benchmarks