--- a/geos/benchmark_test.go
+++ b/geos/benchmark_test.go
@@ -49,3 +49,19 @@ func BenchmarkIntersection(b *testing.B) {
 		})
 	}
 }
+
+func BenchmarkNoOp(b *testing.B) {
+	for _, sz := range []int{10, 100, 1000, 10000} {
+		b.Run(fmt.Sprintf("n=%d", sz), func(b *testing.B) {
+			input := regularPolygon(geom.XY{X: 0, Y: 0}, 1.0, sz).AsGeometry()
+			b.ResetTimer()
+
+			for i := 0; i < b.N; i++ {
+				_, err := noop(input, geom.DisableAllValidations)
+				if err != nil {
+					b.Fatal(err)
+				}
+			}
+		})
+	}
+}